# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API XCRM",
        "description": "Documentação da API do sistema XCRM",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/chamados": {
            "get": {
                "tags": [
                    "Chamados"
                ],
                "summary": "Listar todos os chamados",
                "operationId": "37d9231c3006a7895b5ae87cc9454abb",
                "responses": {
                    "200": {
                        "description": "Lista de chamados retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chamados"
                ],
                "summary": "Criar novo chamado",
                "operationId": "1e1fc54eb55f5b0b96781d05d9b2a372",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clientes_id",
                                    "assunto",
                                    "descricao",
                                    "prioridade",
                                    "status"
                                ],
                                "properties": {
                                    "clientes_id": {
                                        "type": "integer"
                                    },
                                    "assunto": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "prioridade": {
                                        "type": "string",
                                        "enum": [
                                            "baixa",
                                            "media",
                                            "alta"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aberto",
                                            "respondido",
                                            "fechado"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chamado criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chamados/{id}": {
            "get": {
                "tags": [
                    "Chamados"
                ],
                "summary": "Exibir um chamado",
                "operationId": "c9f13efb1d3a091733a8b5c90c096ee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do chamado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chamado encontrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Chamado não encontrado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Chamados"
                ],
                "summary": "Atualizar um chamado",
                "operationId": "7c9f98c7d8a86db8814aead9f279d2fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do chamado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "assunto": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "prioridade": {
                                        "type": "string",
                                        "enum": [
                                            "baixa",
                                            "media",
                                            "alta"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aberto",
                                            "respondido",
                                            "fechado"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chamado atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Chamado não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chamados"
                ],
                "summary": "Remover um chamado",
                "operationId": "ae739a6c789b9c0ea1683be5fca685e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do chamado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chamado removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Chamado não encontrado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Use \"Bearer {token}\" no campo de autorização",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}